Name Updates / PlayerNoah / Replacement-Classes
Made Changes to Names (CamelCase and consistency), added PlayerNoah,
changed replacement-Classes and IF!

+++ src/evaluate/IEvaluator.java
@@ -2,6 +2,6 @@ package evaluate;
 
 import Gameboard.Bitboard;
 
-public interface IEvaluator {
-	public int evaluate(Bitboard gb, long possiblemoves, boolean player);
+public interface IEvaluator{
+    public short evaluate(Bitboard gb, long possiblemoves, boolean player);
 }
